hooge libgbafe updoot: part 2

+++ Tools/libgbafe/gbafe/compress.h
@@ -3,14 +3,14 @@
 
 #include "common.h"
 
-extern uint8_t gGenericBuffer[0x2000]; //! FE8U = 0x2020188
+extern u8 gGenericBuffer[0x2000]; //! FE8U = 0x2020188
 
 #pragma long_calls
 
-void CopyString(char* target, const char* source);                 //! FE8U = 0x8012EC1
-void CopyNoCompData(const void* source, void* target);                 //! FE8U = 0x8012EDD
+void CopyString(char* target, const char* source); //! FE8U = 0x8012EC1
+void CopyNoCompData(const void* source, void* target); //! FE8U = 0x8012EDD
 void Decompress(const void* source, void* target); //! FE8U = 0x8012F51
-size_t GetNoCompDataSize(const void*);                            //! FE8U = 0x8012F91
+unsigned GetNoCompDataSize(const void*); //! FE8U = 0x8012F91
 
 #pragma long_calls_off
